nom_language

Module precedence

Combinators to parse expressions with operator precedence.

Structs

Binary

A binary operator.

LeftAssoc

Parser implementation for the separated_list1 combinator

Unary

An unary operator.

Enums

Assoc

Associativity for binary operators.

Operation

A single evaluation step.

Functions

binary_op

Runs the inner parser and transforms the result into a binary operator with the given precedence and associativity.

left_assoc

Applies a parser multiple times separated by another parser.

precedence

Parses an expression with operator precedence.

unary_op

Runs the inner parser and transforms the result into an unary operator with the given precedence.